+++ app/wishlist/page.tsx
@@ -1,4 +1,3 @@
-import Layout from '@/components/Layout'
 import WishlistManager from '@/components/WishlistManager'
 
 export default function WishlistPage() {